# Conflicts:
#	pom.xml
#	src/main/java/se/iths/projektarbetekomplexjava/ProjektArbeteKomplexJavaApplication.java
#	src/main/java/se/iths/projektarbetekomplexjava/controller/BookController.java
#	src/main/java/se/iths/projektarbetekomplexjava/controller/PublisherController.java
#	src/main/java/se/iths/projektarbetekomplexjava/entity/Customer.java
#	src/main/java/se/iths/projektarbetekomplexjava/entity/Employee.java
#	src/main/java/se/iths/projektarbetekomplexjava/entity/Role.java
#	src/main/java/se/iths/projektarbetekomplexjava/entity/ShoppingCart.java
#	src/main/java/se/iths/projektarbetekomplexjava/repository/BookRepository.java
#	src/main/java/se/iths/projektarbetekomplexjava/repository/RoleRepository.java
#	src/main/java/se/iths/projektarbetekomplexjava/security/CustomerPrincipal.java
#	src/main/java/se/iths/projektarbetekomplexjava/security/EmployeePrincipal.java
#	src/main/java/se/iths/projektarbetekomplexjava/service/AuthorService.java
#	src/main/java/se/iths/projektarbetekomplexjava/service/BookService.java
#	src/main/java/se/iths/projektarbetekomplexjava/service/CustomerService.java
#	src/main/java/se/iths/projektarbetekomplexjava/service/EmployeeService.java
#	src/main/java/se/iths/projektarbetekomplexjava/service/PublisherService.java
#	src/main/java/se/iths/projektarbetekomplexjava/service/StockService.java
#	src/main/resources/application.properties
#	src/main/resources/templates/home.html

+++ src/main/java/se/iths/projektarbetekomplexjava/controller/EmployeeController.java
@@ -5,9 +5,14 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import se.iths.projektarbetekomplexjava.entity.Customer;
 import se.iths.projektarbetekomplexjava.entity.Employee;
+import se.iths.projektarbetekomplexjava.exception.BadRequestException;
+import se.iths.projektarbetekomplexjava.exception.NotAuthorizedException;
+import se.iths.projektarbetekomplexjava.exception.NotFoundException;
+import se.iths.projektarbetekomplexjava.security.PasswordEncoder;
 import se.iths.projektarbetekomplexjava.service.CustomerService;
 import se.iths.projektarbetekomplexjava.service.EmployeeService;
 
+import java.util.List;
 import java.util.Optional;
 
 @RestController
@@ -16,46 +21,89 @@ public class EmployeeController {
 
     private final CustomerService customerService;
     private final EmployeeService employeeService;
+    private PasswordEncoder passwordEncoder;
 
-    public EmployeeController(CustomerService customerService, EmployeeService employeeService) {
+    public EmployeeController(CustomerService customerService, EmployeeService employeeService, PasswordEncoder passwordEncoder) {
         this.customerService = customerService;
         this.employeeService = employeeService;
+        this.passwordEncoder = passwordEncoder;
     }
 
     @PostMapping("/signup")
     public ResponseEntity<Employee> addEmployee(@RequestBody Employee employee){
         Employee addedEmployee = employeeService.addEmployee(employee);
+        if (addedEmployee.getFirstName().isEmpty() || addedEmployee.getLastName().isEmpty() || addedEmployee.getAddress().isEmpty()
+                || addedEmployee.getPhone().isEmpty() || addedEmployee.getUsername().isEmpty() || addedEmployee.getEmail().isEmpty() || addedEmployee.getPassword().isEmpty()){
+            throw new BadRequestException("Insufficient data, please fill the required registration data.");
+        }
         return new ResponseEntity<>(addedEmployee, HttpStatus.CREATED);
     }
 
-    @GetMapping("/login")
-    public ResponseEntity<Employee> logInEmployee(@RequestBody Employee employee){
-        Employee loginEmployee = employeeService.getEmployeeByUsername(employee.getUsername(), employee.getPassword());
-        return new ResponseEntity<>(loginEmployee, HttpStatus.OK);
+    @PostMapping("/login")
+    public Object logInCustomer(@RequestBody Employee employee){
+        Optional<Employee> loginEmployee = employeeService.getEmployeeByEmail(employee.getEmail(), employee.getPassword());
+        List<Employee> employeeList = employeeService.getByEmail(employee.getEmail());
+        for (Employee employees:employeeList){
+            if (passwordEncoder.bCryptPasswordEncoder().matches(employee.getPassword(), employees.getPassword())){
+                return employeeList;
+            }
+        }
+        return loginEmployee.orElseThrow(() -> new NotAuthorizedException("Invalid login, please enter right login data or create new account"));
+    }
+
+    @PutMapping("/update")
+    public ResponseEntity<Employee> updateEmployee(@RequestBody Employee employee){
+        Employee updatedEmployee = employeeService.updateEmployee(employee);
+        if (updatedEmployee == null){
+            throw new NotFoundException("User is not in the system.");
+        }
+        return new ResponseEntity<>(updatedEmployee, HttpStatus.CREATED);
+    }
+
+    @DeleteMapping("/deleteCustomer/{id}")
+    public ResponseEntity<Void> removeCustomer(@PathVariable Long id){
+        Optional<Employee> foundEmployee = employeeService.findUserById(id);
+        if (foundEmployee.isEmpty()){
+            throw new NotFoundException("No data available of user ID: " + id);
+        }
+        employeeService.removeEmployee(id);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    }
+
+
+    @GetMapping("/getListOfCustomer")
+    public ResponseEntity<Iterable<Customer>> getAllCustomers(){
+        List<Customer> allCustomers = employeeService.findAllCustomers();
+        if (allCustomers.isEmpty()){
+            throw new NotFoundException("No data available in the system.");
+        }
+        return new ResponseEntity<>(allCustomers, HttpStatus.OK);
+    }
+
+    @GetMapping("/getListOfEmployee")
+    public ResponseEntity<Iterable<Employee>> getAllEmployees(){
+        List<Employee> allEmployees = employeeService.findAllEmployees();
+        if (allEmployees.isEmpty()){
+            throw new NotFoundException("No data available in the system.");
+        }
+        return new ResponseEntity<>(allEmployees, HttpStatus.OK);
     }
 
     @GetMapping("/searchEmployee/{id}")
     public ResponseEntity<Optional<Employee>> findEmployeeId(@PathVariable long id){
         Optional<Employee> foundEmployee = employeeService.findUserById(id);
+        if (foundEmployee.isEmpty()){
+            throw new NotFoundException("No data available of user ID: " + id);
+        }
         return new ResponseEntity<>(foundEmployee, HttpStatus.OK);
-
     }
 
     @GetMapping("/searchCustomer/{id}")
     public ResponseEntity<Optional<Customer>> findCustomerId(@PathVariable long id){
         Optional<Customer> foundCustomer = customerService.findUserById(id);
+        if (foundCustomer.isEmpty()){
+            throw new NotFoundException("No data available of user ID: " + id);
+        }
         return new ResponseEntity<>(foundCustomer, HttpStatus.OK);
-    }
-
-    @GetMapping("/getListOfCustomer")
-    public ResponseEntity<Iterable<Customer>> getAllCustomers(){
-        Iterable<Customer> allCustomers = employeeService.findAllCustomers();
-        return new ResponseEntity<>(allCustomers, HttpStatus.OK);
-    }
-
-    @DeleteMapping("/deleteEmployee/{id}")
-    public ResponseEntity<Void> removeEmployee(@PathVariable Long id){
-        employeeService.removeEmployee(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
